Move the version cmd to its own file
