+++ backend/cmd/version.go
@@ -1,0 +1,28 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/sensu/sensu-go/version"
+	"github.com/spf13/cobra"
+)
+
+func init() {
+	rootCmd.AddCommand(newVersionCommand())
+}
+
+func newVersionCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "version",
+		Short: "Show the sensu-backend version information",
+		Run: func(cmd *cobra.Command, args []string) {
+			fmt.Printf("sensu-backend version %s, build %s, built %s\n",
+				version.Semver(),
+				version.BuildSHA,
+				version.BuildDate,
+			)
+		},
+	}
+
+	return cmd
+}
